Refactor to match the common use
